+++ src/main/kotlin/gomushin/backend/couple/domain/service/CoupleService.kt
@@ -38,6 +38,6 @@ class CoupleService(
 
     @Transactional
     fun deleteByMemberId(memberId: Long) {
-        return coupleRepository.deleteByMemberId(memberId)
+        coupleRepository.deleteByMemberId(memberId)
     }
 }
