+++ src/main/kotlin/gomushin/backend/member/domain/service/MemberService.kt
@@ -46,6 +46,6 @@ class MemberService(
 
     @Transactional
     fun deleteMember(id : Long) {
-        return memberRepository.deleteById(id)
+        memberRepository.deleteById(id)
     }
 }
